+++ src/main/java/forestry/arboriculture/render/RenderFruitPodBlock.java
@@ -54,10 +54,10 @@ public class RenderFruitPodBlock implements ISimpleBlockRenderingHandler {
 		double d1 = 15.0D;
 		double d2 = 4.0D;
 		double d3 = 4.0D + l1;
-		double d4 = (double) podIcon.getInterpolatedU(d0);
-		double d5 = (double) podIcon.getInterpolatedU(d1);
-		double d6 = (double) podIcon.getInterpolatedV(d2);
-		double d7 = (double) podIcon.getInterpolatedV(d3);
+		double d4 = podIcon.getInterpolatedU(d0);
+		double d5 = podIcon.getInterpolatedU(d1);
+		double d6 = podIcon.getInterpolatedV(d2);
+		double d7 = podIcon.getInterpolatedV(d3);
 		double d8 = 0.0D;
 		double d9 = 0.0D;
 
@@ -107,10 +107,10 @@ public class RenderFruitPodBlock implements ISimpleBlockRenderingHandler {
 			i2 = k1 - 1;
 		}
 
-		d4 = (double) podIcon.getMinU();
-		d5 = (double) podIcon.getInterpolatedU((double) i2);
-		d6 = (double) podIcon.getMinV();
-		d7 = (double) podIcon.getInterpolatedV((double) i2);
+		d4 = podIcon.getMinU();
+		d5 = podIcon.getInterpolatedU(i2);
+		d6 = podIcon.getMinV();
+		d7 = podIcon.getInterpolatedV(i2);
 
 		tessellator.addVertexWithUV(d10, d13, d15, d4, d7);
 		tessellator.addVertexWithUV(d11, d13, d15, d5, d7);
@@ -120,10 +120,10 @@ public class RenderFruitPodBlock implements ISimpleBlockRenderingHandler {
 		tessellator.addVertexWithUV(d11, d12, d14, d5, d6);
 		tessellator.addVertexWithUV(d11, d12, d15, d5, d7);
 		tessellator.addVertexWithUV(d10, d12, d15, d4, d7);
-		d4 = (double) podIcon.getInterpolatedU(12.0D);
-		d5 = (double) podIcon.getMaxU();
-		d6 = (double) podIcon.getMinV();
-		d7 = (double) podIcon.getInterpolatedV(4.0D);
+		d4 = podIcon.getInterpolatedU(12.0D);
+		d5 = podIcon.getMaxU();
+		d6 = podIcon.getMinV();
+		d7 = podIcon.getInterpolatedV(4.0D);
 		d8 = 8.0D;
 		d9 = 0.0D;
 		double d16;
